Big clean up of condition/action logic
