+++ apps/web/utils/thread.ts
@@ -1,4 +1,4 @@
 // The first message id in a thread is the threadId
-export function isReplyInThread(messageId: string, threadId: string) {
-  return messageId !== threadId;
+export function isReplyInThread(messageId: string, threadId: string): boolean {
+  return !!(messageId && messageId !== threadId);
 }
